reorganize MovieCard component
